+++ pkg/helm/chart.go
@@ -16,13 +16,12 @@ func (chart *Chart) Version() semver.Version {
 	return version
 }
 
-func (helmClient *Client) GetLatestChart(name, repoUrl string) (*Chart, error) {
+func (helmClient *Client) GetLatestChart(name string) (*Chart, error) {
 	var err error
 	var chartPath string
 	var chart *chart.Chart
 
 	client := action.NewShowWithConfig(action.ShowChart, helmClient.cfg)
-	client.RepoURL = repoUrl
 
 	if chartPath, err = client.ChartPathOptions.LocateChart(name, helmClient.settings); err != nil {
 		return nil, err